+++ aiagent/src/components/BaseTemplate.tsx
